build: migrate to multi-module Maven and fix integration test lifecycle
Add root pom.xml (Spring Boot BOM parent, packaging=pom) and wire
backend/pom.xml as a child module. Foundation for adding the frontend
and performance-testing modules in later sprints.

Fix a test reliability bug in AbstractIntegrationTest: @Testcontainers
+ @Container gave containers a per-test-class lifecycle. With Spring's
test-context cache (one shared ApplicationContext across all classes),
the @AfterAll podman rm -f swept containers away while the cached context
still held open HikariCP and AMQP pools — causing CONNECTION_FORCED
shutdowns and cascading Connection refused errors across the suite.

Switch to the Testcontainers singleton pattern: containers start once in
a static block and live for the full JVM lifetime, matching the context
cache. Orphaned containers from previous aborted runs are purged by a
labelled podman rm -f sweep before start; a JVM shutdown hook repeats
the sweep at run end (Ryuk is disabled in this rootless-Podman env).

Also:
- V1 migration: DATETIME2 -> DATETIMEOFFSET(6) for timezone-aware timestamps
- WardOverviewRecord: activeProblemCount long -> Integer (nullable for patients with no problems)
- EhrControllerTest: idempotent EHR creation now correctly asserts 409 Conflict
- Add devops/infra/dev/scripts/cleanup-testcontainers.sh helper

+++ backend/src/main/java/com/arcticsurge/cosmolab/infrastructure/persistence/WardOverviewRecord.java
@@ -66,5 +66,5 @@ public class WardOverviewRecord {
     private BigDecimal weight;
 
     @Column(name = "active_problem_count")
-    private long activeProblemCount;
+    private Integer activeProblemCount;
 }
